requirements.txt
prefect>=2.18.0
mlflow>=2.9.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
yfinance>=0.2.30
SQLAlchemy>=2.0.0
python-dotenv>=0.1.0
pytz>=2023.3
joblib>=1.3.0
tensorflow>=2.10.0
mlflow-dagshub>=0.1.0
streamlit>=1.28.0
plotly>=5.17.0